requirements.txt
pywin32
easydict
PySide6
PySide6-Fluent-Widgets
fusepy
diskcache
requests
pyyaml